+++ package.json
@@ -16,6 +16,7 @@
     "@vanilla-extract/css": "^1.11.0",
     "@vanilla-extract/dynamic": "^2.0.3",
     "axios": "^1.4.0",
+    "cookie": "^0.5.0",
     "fast-diff": "^1.3.0",
     "googleapis": "^118.0.0",
     "jwt-decode": "^3.1.2",
@@ -27,6 +28,7 @@
   },
   "devDependencies": {
     "@tanstack/react-query-devtools": "^4.29.12",
+    "@types/cookie": "^0.5.1",
     "@types/node": "^18.16.0",
     "@types/react": "^18.2.0",
     "@types/react-dom": "^18.2.0",
