+++ src/pages/index.tsx
@@ -1,5 +1,6 @@
 import Head from 'next/head';
 import * as style from '@/styles/style-test.css.ts';
+import { getAccessTokenByRefreshToken } from '@/features/auth/apis/index.ts';
 
 export default function Home() {
   return (
@@ -7,7 +8,7 @@ export default function Home() {
       <Head>
         <title>홈 입니다</title>
       </Head>
-
+      <button onClick={getAccessTokenByRefreshToken}>get refresh token!</button>
       <div className={style.Main}>Home</div>
     </>
   );
