+++ tsconfig.json
@@ -34,6 +34,5 @@
   "exclude": [
     "node_modules",
     ".next",
-    "src/pages/api/auth.ts"
   ],
 }
